+++ scripts/generate.js
@@ -1,35 +1,6 @@
-import { readFile, writeFile, rm } from 'node:fs/promises';
 import path from 'node:path';
+import { getEventCatalogConfigFile, cleanup } from './eventcatalog-config-file-utils.js';
 
-/**
- * Very strange behaviour when importing ESM files from catalogs into core.
- * Core (node) does not know how to handle ESM files, so we have to try and convert them.
- *
- * This needs sorting out! Sorry if you are reading this, but it unblocked me for now!
- * @param {*} content
- * @returns
- */
-function convertESMtoCJS(content) {
-  // Replace import statements with require
-  content = content.replace(/import\s+([a-zA-Z0-9{},\s*]+)\s+from\s+['"]([^'"]+)['"];/g, (match, imports, modulePath) => {
-    return `const ${imports.trim()} = require('${modulePath}');`;
-  });
-
-  // Replace export default with module.exports
-  content = content.replace(/export\s+default\s+/g, 'module.exports = ');
-
-  // Replace named exports with module.exports
-  content = content.replace(/export\s+{([^}]+)}/g, (match, exports) => {
-    return `module.exports = {${exports.trim()}};`;
-  });
-
-  // Remove declarations of __filename and __dirname
-  content = content.replace(/^\s*(const|let|var)\s+__(filename|dirname)\s+=\s+.*;?\s*$/gm, '');
-
-  return content;
-}
-
-// TODO: Do we actually need this? Can we clean this up
 function getDefaultExport(importedModule) {
   if (importedModule === null || typeof importedModule !== 'object') {
     throw new Error('Invalid module');
@@ -46,21 +17,10 @@ function getDefaultExport(importedModule) {
   return importedModule;
 }
 
-async function cleanup() {
-  await rm(path.join(process.env.PROJECT_DIR, 'eventcatalog.config.cjs'));
-}
-
 const generate = async () => {
+  const PROJECT_DIRECTORY = process.env.PROJECT_DIR;
   try {
-    // Fix for the file
-    const rawFile = await readFile(path.join(process.env.PROJECT_DIR, 'eventcatalog.config.js'), 'utf8');
-
-    // Have to conver the ESM to CJS...
-    const configAsCommonJS = convertESMtoCJS(rawFile);
-    await writeFile(path.join(process.env.PROJECT_DIR, 'eventcatalog.config.cjs'), configAsCommonJS);
-
-    const configAsCJS = await import(path.join(process.env.PROJECT_DIR, 'eventcatalog.config.cjs'));
-    const config = configAsCJS.default;
+    const config = await getEventCatalogConfigFile(PROJECT_DIRECTORY);
 
     const { generators = [] } = config;
 
@@ -69,19 +29,16 @@ const generate = async () => {
       return;
     }
 
-    // Tidy up
-    await writeFile(path.join(process.env.PROJECT_DIR, 'eventcatalog.config.js'), rawFile);
-
     for (const generator of generators) {
       let plugin = generator[0];
       const pluginConfig = generator[1];
 
       if (plugin.startsWith('./')) {
-        plugin = path.join(process.env.PROJECT_DIR, plugin);
+        plugin = path.join(PROJECT_DIRECTORY, plugin);
       }
 
       if (plugin.includes('<rootDir>')) {
-        plugin = plugin.replace('<rootDir>', process.env.PROJECT_DIR);
+        plugin = plugin.replace('<rootDir>', PROJECT_DIRECTORY);
       }
 
       try {
@@ -95,14 +52,16 @@ const generate = async () => {
         // Use importedGenerator here
       } catch (error) {
         console.error('Error loading plugin:', error);
-        await cleanup();
+        await cleanup(PROJECT_DIRECTORY);
         return;
       }
     }
+
+    await cleanup(PROJECT_DIRECTORY);
   } catch (error) {
     // Failed to generate clean up...
     console.error(error);
-    await cleanup();
+    await cleanup(PROJECT_DIRECTORY);
   }
 };