+++ src/components/MDX/components.tsx
@@ -4,8 +4,13 @@ import File from '@components/MDX/File';
 import Accordion from '@components/MDX/Accordion/Accordion.astro';
 import AccordionGroup from '@components/MDX/Accordion/AccordionGroup.astro';
 import Flow from '@components/MDX/Flow/Flow.astro';
+import Tiles from '@components/MDX/Tiles/Tiles.astro';
+import Tile from '@components/MDX/Tiles/Tile.astro';
+import Steps from '@components/MDX/Steps/Steps.astro';
+import Step from '@components/MDX/Steps/Step.astro';
 import Admonition from '@components/MDX/Admonition';
-import OpenAPI from '@components/MDX/OpenAPI/OpenAPI';
+import OpenAPI from '@components/MDX/OpenAPI/OpenAPI.astro';
+import AsyncAPI from '@components/MDX/AsyncAPI/AsyncAPI.astro';
 
 //  Portals: required for server/client components
 import NodeGraphPortal from '@components/MDX/NodeGraph/NodeGraphPortal';
@@ -16,11 +21,16 @@ const components = (props: any) => {
     Accordion,
     AccordionGroup,
     Flow,
+    OpenAPI,
+    AsyncAPI,
+    Tile,
+    Tiles,
+    Step,
+    Steps,
     Admonition: (mdxProp: any) => <Admonition {...mdxProp} {...props} />,
     File: (mdxProp: any) => File({ ...props, ...mdxProp }),
     NodeGraph: (mdxProp: any) => NodeGraphPortal({ ...props.data, ...mdxProp }),
     SchemaViewer: (mdxProp: any) => SchemaViewerPortal({ ...props.data, ...mdxProp }),
-    OpenAPI: (mdxProp: any) => <OpenAPI {...mdxProp} {...props} />,
     Schema: (mdxProp: any) => Schema({ ...props, ...mdxProp }),
   };
 };