+++ src/utils/messages.ts
@@ -1,7 +1,13 @@
-// // import type { Domain, Message } from '@eventcatalog/types';
-// import { getCollection,  } from 'astro:content';
+// Exporting getCommands and getEvents directly
+import { getCommands } from '@utils/commands';
+import { getEvents } from '@utils/events';
+export { getCommands } from '@utils/commands';
+export { getEvents } from '@utils/events';
 
-// export const getMessages = async (): Promise<Message[]> => {
-//   const commands = await getCollection('commands');
-//   return commands.map((message: any) => message.data);
-// };
+// Main function that uses the imported functions
+export const getMessages = async () => {
+  const commands = await getCommands();
+  const events = await getEvents();
+
+  return { commands, events };
+};